Fixed a security bug, and wrote a test to prevent it from happening again.
There was a bug in the GnuPG.list_all_keyids function. This function is supposed to return a list of keyids for the specific fingerprint that's passed into it. Instead, it returned the list of keyids for _all keys_ in the keyring. The GnuPG.verify function first verifies that the signature is valid, then confirms that the signature was signed by the right key using the list_all_keyids function and passing in the authority key fingerprint. But because the list_all_keyids function had this bug, the verify function treated any signature as successful as long as the signing key was in the keyring.

GPG Sync creates a new temporary keyring to download keys into. This means that, so long as the fingerprints file is signed by any key in this keyring (essentially, any key listed on the fingerprints file -- or, someone in your organization), the signature would be accepted as valid. Instead, it should only be accepted as valid if it's signed by the authority key.